update spec tests
* Parse memory pages as uint64, as we need to check whether it's invalid
  in assert_invalid, and one of the spec tests uses the value 2**32.
* Update WASM_MAX_PAGES to 65535 (not 65536)
* Add the new spec tests: binary.wast, break-drop.wast and
  typecheck.wast. A couple of these were added before, but weren't added
  as .txt files in the test directory.

+++ src/wasm-ast.h
@@ -243,8 +243,8 @@ WASM_DEFINE_VECTOR(segment, WasmSegment);
 
 typedef struct WasmMemory {
   WasmLocation loc;
-  uint32_t initial_pages;
-  uint32_t max_pages;
+  uint64_t initial_pages;
+  uint64_t max_pages;
   WasmSegmentVector segments;
 } WasmMemory;
